Fix: organized the code.
