+++ ios/Classes/SwiftFlutterBluetoothSerialPlugin.swift
@@ -2,14 +2,6 @@ import Flutter
 import UIKit
 import CoreBluetooth
 
-extension CBUUID {
-    var fullUUIDString: String {
-        uuidString.count == 4
-            ? String(format: "0000%@-0000-1000-8000-00805F9B34FB", uuidString)
-            : uuidString.lowercased()
-    }
-}
-
 enum FlutterMethodName: String {
     case isAvailable
     case isOn
@@ -37,34 +29,39 @@ enum FlutterMethodName: String {
     case write
 }
 
-private let stateEventChannelStreamHandler = StateEventChannelStreamHandler()
-private let discveryEventChannelStreamHandler = DiscoveryEventChannelStreamHandler()
+enum FlutterChannelName: String {
+    case bluetoothSerialMethod = "flutter_bluetooth_serial/methods"
+    case bluetoothSerialState = "flutter_bluetooth_serial/state"
+    case bluetoothSerialDiscovery = "flutter_bluetooth_serial/discovery"
+}
 
 public class SwiftFlutterBluetoothSerialPlugin: NSObject, FlutterPlugin {
+    private static let stateEventChannelStreamHandler = StateEventChannelStreamHandler()
+    private static let discveryEventChannelStreamHandler = DiscoveryEventChannelStreamHandler()
+
     private var bluetoothManager: BluetoothManager?
-    private var isInitialized = false
 
     public static func register(with registrar: FlutterPluginRegistrar) {
         let messenger = registrar.messenger()
-        let channel = FlutterMethodChannel(name: "flutter_bluetooth_serial/methods", binaryMessenger: messenger)
+        let channel = FlutterMethodChannel(name: FlutterChannelName.bluetoothSerialMethod.rawValue, binaryMessenger: messenger)
         let instance = SwiftFlutterBluetoothSerialPlugin()
         registrar.addMethodCallDelegate(instance, channel: channel)
 
-        let stateChannel = FlutterEventChannel(name: "flutter_bluetooth_serial/state", binaryMessenger: messenger)
+        let stateChannel = FlutterEventChannel(name: FlutterChannelName.bluetoothSerialState.rawValue, binaryMessenger: messenger)
         stateChannel.setStreamHandler(stateEventChannelStreamHandler)
 
-        let discoveryChannel = FlutterEventChannel(name: "flutter_bluetooth_serial/discovery", binaryMessenger: messenger)
+        let discoveryChannel = FlutterEventChannel(name: FlutterChannelName.bluetoothSerialDiscovery.rawValue, binaryMessenger: messenger)
         discoveryChannel.setStreamHandler(discveryEventChannelStreamHandler)
     }
 
     public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
-        print("FlutterBluetoothSerial: handle called: ", call.method)
         if bluetoothManager == nil {
-            bluetoothManager = BluetoothManager()
+            bluetoothManager = BluetoothManager(stateEventChannelStreamHandler: SwiftFlutterBluetoothSerialPlugin.stateEventChannelStreamHandler,
+                                                discveryEventChannelStreamHandler: SwiftFlutterBluetoothSerialPlugin.discveryEventChannelStreamHandler)
         }
 
         guard let method = FlutterMethodName(rawValue: call.method) else {
-            result("unknown method '\(call.method)' called.")
+            result(FlutterError(code: "-1", message: "unknown method '\(call.method)' called.", details: nil))
             return
         }
 
@@ -104,7 +101,7 @@ public class SwiftFlutterBluetoothSerialPlugin: NSObject, FlutterPlugin {
 
         case .setName:
             print("\(method.rawValue) called.")
-            result(FlutterError(code: "", message: "It can't modify device name from program in iOS", details: nil))
+            result(FlutterError(code: "-1", message: "It can't modify device name from App in iOS", details: nil))
 
         case .getDeviceBondState:
             print("\(method.rawValue) called.")
@@ -162,146 +159,3 @@ public class SwiftFlutterBluetoothSerialPlugin: NSObject, FlutterPlugin {
         }
     }
 }
-
-class BluetoothManager: NSObject {
-    var isAvailable: Bool {
-        centralManager?.state != .unsupported && centralManager?.state != .unknown
-    }
-    var isOn: Bool {
-        centralManager?.state == .poweredOn
-    }
-    var mangerStateInt: Int {
-        centralManager.map { self.convertToStateNumber(from: $0.state) } ?? -1
-    }
-    var isDiscovering: Bool {
-        centralManager?.isScanning ?? false
-    }
-
-    private var centralManager: CBCentralManager?
-    private var scannedPeripherals: [(peripheral: CBPeripheral, rssi: Int)] = []
-
-    override init() {
-        super.init()
-
-        centralManager = CBCentralManager(delegate: self, queue: nil)
-        print("manager state: \(String(describing: centralManager?.state))")
-    }
-
-    func connect(with argument: Any?) throws {
-//        let data = FlutterStandardDataType(rawValue: argument)
-    }
-
-    func startDiscovery() {
-        centralManager?.scanForPeripherals(withServices: nil)
-    }
-
-    func cancelDiscovery() {
-        centralManager?.stopScan()
-    }
-
-    func bondedDevice() -> [[String: Any]] {
-        scannedPeripherals.filter { $0.peripheral.state == .connected }.map { self.convertToDictionary(from: $0.peripheral, rssi: $0.rssi) }
-    }
-
-    func deviceBondState(at identifier: String) -> Int {
-        scannedPeripherals
-            .first { $0.peripheral.identifier.uuidString == identifier }
-            .map { self.convertToConnectedNumber(from: $0.peripheral.state) } ?? 0
-    }
-}
-
-extension BluetoothManager: CBCentralManagerDelegate {
-    func centralManagerDidUpdateState(_ central: CBCentralManager) {
-        print("new state: \(central.state)")
-        stateEventChannelStreamHandler.sink(stateInteger: convertToStateNumber(from: central.state))
-    }
-
-    func centralManager(_ central: CBCentralManager, didDiscover peripheral: CBPeripheral, advertisementData: [String : Any], rssi RSSI: NSNumber) {
-        scannedPeripherals.append((peripheral: peripheral, rssi: RSSI.intValue))
-        discveryEventChannelStreamHandler.sink(argument: convertToDictionary(from: peripheral, rssi: RSSI.intValue))
-    }
-}
-
-private extension BluetoothManager {
-    func convertToStateNumber(from state: CBManagerState) -> Int {
-        switch state {
-        case .unknown:
-            return state.rawValue
-        case .unsupported, .unauthorized, .resetting: // TODO: set "right" number
-            return 0
-        case .poweredOff:
-            return 10
-        case .poweredOn:
-            return 12
-        }
-    }
-
-    func convertToDictionary(from peripheral: CBPeripheral, rssi: Int) -> [String: Any] {
-        let name = peripheral.name ?? ""
-        let address = peripheral.identifier.uuidString
-        let type = 0
-        let isConnected = false // TODO: It need calculate
-        let bondState = convertToConnectedNumber(from: peripheral.state)
-
-        return [
-            "name": name,
-            "address": address,
-            "type": type,
-            "isConnected": isConnected,
-            "bondState": bondState,
-            "rssi": rssi
-        ]
-    }
-
-    func convertToConnectedNumber(from state: CBPeripheralState) -> Int {
-        switch state {
-        case .disconnected, .disconnecting:
-            return 10
-        case .connecting:
-            return 11
-        case .connected:
-            return 12
-        }
-    }
-}
-
-class StateEventChannelStreamHandler: NSObject, FlutterStreamHandler {
-    private(set) var stateSink: FlutterEventSink?
-
-    public func sink(stateInteger: Int) {
-        stateSink?(stateInteger)
-    }
-
-    public func onListen(withArguments arguments: Any?, eventSink events: @escaping FlutterEventSink) -> FlutterError? {
-        stateSink = events
-
-        return nil
-    }
-
-    public func onCancel(withArguments arguments: Any?) -> FlutterError? {
-        stateSink = nil
-
-        return nil
-    }
-}
-
-class DiscoveryEventChannelStreamHandler: NSObject, FlutterStreamHandler {
-    private(set) var discoverySink: FlutterEventSink?
-
-    public func sink(argument: Any?) {
-        discoverySink?(argument)
-    }
-
-    public func onListen(withArguments arguments: Any?, eventSink events: @escaping FlutterEventSink) -> FlutterError? {
-        discoverySink = events
-
-        return nil
-    }
-
-    public func onCancel(withArguments arguments: Any?) -> FlutterError? {
-        discoverySink = nil
-
-        return nil
-    }
-}
-
